Move provider connection popup into a reusable hook
